requirements.txt
python-telegram-bot==21.0
web3==6.15.1
eth-account==0.11.0
